refactor: DDDの設計思考に沿ってディレクトリ構成を修正

+++ backend/domain/src/main/kotlin/com/example/daycare/domain/Student.kt
@@ -1,8 +1,8 @@
 package com.example.daycare.domain
 
 data class Student(
-    val id: Int,
+    val id: Long,
     val name: String,
     val age: Int,
     val className: String
-)
+)
